+++ TripPaw_React/pages/admin/index.js
@@ -1,0 +1,11 @@
+import AdminpageLayout from "@/components/layout/AdminpageLayout";
+import React from "react";
+
+const Adminpage = () => {
+  return (
+    <AdminpageLayout>
+
+    </AdminpageLayout>);
+}
+
+export default Adminpage;
